requirements.txt
# Tapis
tapipy
tapisservice==1.2.2

# Service API
pydantic
fastapi
sqlmodel
uvicorn

# Databases/APIs
sqlalchemy
alembic
psycopg2
kubernetes
neo4j-driver
rabbitpy
channelpy

# Misc
pylint

# Dev
jupyterlab
